Input: Implemented mouse-based input devices (Windows only)

+++ Core/NES/Loaders/iNesLoader.cpp
@@ -83,7 +83,7 @@ void iNesLoader::LoadRom(RomData& romData, vector<uint8_t>& romFile, NesHeader *
 	romData.PrgRom.insert(romData.PrgRom.end(), buffer, buffer + prgSize);
 	while(romData.PrgRom.size() < 256) {
 		//Ensure the PRG is at least 256 bytes in size by mirroring as needed (mapper code requires it to be 256 bytes at least)
-		romData.PrgRom.insert(romData.PrgRom.end(), buffer, buffer + std::min<int>(prgSize, 256 - romData.PrgRom.size()));
+		romData.PrgRom.insert(romData.PrgRom.end(), buffer, buffer + std::min<int>(prgSize, 256 - (int)romData.PrgRom.size()));
 	}
 	buffer += prgSize;
 	romData.ChrRom.insert(romData.ChrRom.end(), buffer, buffer + chrSize);
